requirements.txt
pandas
numpy
scikit-learn
scipy
yfinance
fredapi
requests
python-dotenv
tensorflow>=2.13
transformers
torch
vaderSentiment
fastapi
uvicorn[standard]
dash
plotly
shap
joblib
beautifulsoup4
lxml
